combined LoginDTO and CreateAccountDTO
