+++ src/main/java/org/launchcode/brewpub/models/dto/CreateAccountDTO.java
@@ -2,11 +2,22 @@ package org.launchcode.brewpub.models.dto;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
-public class CreateAccountDTO extends LoginDTO {
-    private String verifyPassword;
+public class CreateAccountDTO {
 
+    @NotNull
+    @NotBlank
+    @Size(min = 5, max =40 , message = "Invalid username. Must be between 5 and 40 characters.")
+    private String username;
+
+    @NotNull
+    @NotBlank
+    @Size(min = 7, max = 30, message = "Invalid password. Must be between 7 and 30 characters.")
+    private String password;
+
+    private String verifyPassword;
 
     @NotBlank
     @Size(min =2, max = 40, message = "First name must be between 2 and 40 characters")
@@ -20,6 +31,22 @@ public class CreateAccountDTO extends LoginDTO {
     @Email(message = "Please enter a valid email")
     @Size(max = 120)
     private String email;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
 
     public String getFirstName() {
         return firstName;
